+++ src/commands/index.ts
@@ -1,1 +1,2 @@
 export { fetchCommand } from "./fetch";
+export { storeCommand } from "./store";